feat(meta): add client subdir for lib
