+++ tailwind.config.mjs
@@ -7,7 +7,11 @@ export default {
         "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}",
     ],
     theme: {
-        extend: {},
+        extend: {
+            spacing: {
+                terminal: "80ch"
+            }
+        },
     },
     darkMode: "class",
     plugins: [nextui()],
